feat(actor sheets): generalize shared actor sheet components
