+++ src/system/applications/actor/base.ts
@@ -2,9 +2,33 @@ import { Resource } from '@src/system/types/cosmere';
 import { CosmereActor } from '@system/documents/actor';
 import { DeepPartial } from '@system/types/utils';
 
+// Mixins
+import {
+    TabsApplicationMixin,
+    DragDropApplicationMixin,
+    ComponentHandlebarsApplicationMixin,
+} from '@system/applications/mixins';
+
+// Components
+import {
+    ActorDetailsComponent,
+    ActorResourceComponent,
+    ActorAttributesComponent,
+    ActorActionsListComponent,
+    ActorSearchBarComponent,
+    ActorConditionsComponent,
+    ActorInjuriesListComponent,
+} from './components';
+
 const { ActorSheetV2 } = foundry.applications.sheets;
 
 export type ActorSheetMode = 'view' | 'edit';
+
+export const enum BaseSheetTab {
+    Actions = 'actions',
+    Equipment = 'equipment',
+    Effects = 'effects',
+}
 
 // NOTE: Have to use type instead of interface to comply with AnyObject type
 // eslint-disable-next-line @typescript-eslint/consistent-type-definitions
@@ -14,7 +38,9 @@ export type BaseActorSheetRenderContext = {
 
 export class BaseActorSheet<
     T extends BaseActorSheetRenderContext = BaseActorSheetRenderContext,
-> extends ActorSheetV2<T> {
+> extends TabsApplicationMixin(
+    DragDropApplicationMixin(ComponentHandlebarsApplicationMixin(ActorSheetV2)),
+)<T> {
     /* eslint-disable @typescript-eslint/unbound-method */
     static DEFAULT_OPTIONS = foundry.utils.mergeObject(
         foundry.utils.mergeObject({}, super.DEFAULT_OPTIONS),
@@ -26,9 +52,47 @@ export class BaseActorSheet<
                 handler: this.onFormEvent,
                 submitOnChange: true,
             } as unknown,
+            dragDrop: [
+                {
+                    dragSelector: '[data-drag]',
+                    dropSelector: '*',
+                },
+            ],
         },
     );
     /* eslint-enable @typescript-eslint/unbound-method */
+
+    static COMPONENTS = foundry.utils.mergeObject(super.COMPONENTS, {
+        'app-actor-details': ActorDetailsComponent,
+        'app-actor-resource': ActorResourceComponent,
+        'app-actor-attributes': ActorAttributesComponent,
+        'app-actor-actions-list': ActorActionsListComponent,
+        'app-actor-search-bar': ActorSearchBarComponent,
+        'app-actor-conditions': ActorConditionsComponent,
+        'app-actor-injuries-list': ActorInjuriesListComponent,
+    });
+
+    static PARTS = foundry.utils.mergeObject(super.PARTS, {
+        navigation: {
+            template:
+                'systems/cosmere-rpg/templates/actors/parts/navigation.hbs',
+        },
+    });
+
+    static TABS = foundry.utils.mergeObject(super.TABS, {
+        [BaseSheetTab.Actions]: {
+            label: 'COSMERE.Actor.Sheet.Tabs.Actions',
+            icon: '<i class="cosmere-icon">3</i>',
+        },
+        [BaseSheetTab.Equipment]: {
+            label: 'COSMERE.Actor.Sheet.Tabs.Equipment',
+            icon: '<i class="fa-solid fa-suitcase"></i>',
+        },
+        [BaseSheetTab.Effects]: {
+            label: 'COSMERE.Actor.Sheet.Tabs.Effects',
+            icon: '<i class="fa-solid fa-bolt"></i>',
+        },
+    });
 
     get actor(): CosmereActor {
         return super.document;
@@ -42,9 +106,13 @@ export class BaseActorSheet<
 
     /* --- Actions --- */
 
-    public static async onToggleMode(this: BaseActorSheet, _: Event) {
-        // this.mode = this.mode === 'view' ? 'edit' : 'view';
+    public static async onToggleMode(this: BaseActorSheet, event: Event) {
+        if (!(event.target instanceof HTMLInputElement)) return;
 
+        // Stop event propagation
+        event.stopPropagation();
+
+        // Update the actor
         await this.actor.update(
             {
                 'flags.cosmere-rpg.sheetMode':
@@ -53,7 +121,22 @@ export class BaseActorSheet<
             { render: false },
         );
 
+        // Render the sheet
         void this.render(true);
+
+        // Get toggle
+        const toggle = $(this.element).find('#mode-toggle');
+
+        // Update checked status
+        toggle.find('input').prop('checked', this.mode === 'edit');
+
+        // Update tooltip
+        toggle.attr(
+            'data-tooltip',
+            game.i18n!.localize(
+                `COSMERE.Actor.Sheet.${this.mode === 'edit' ? 'View' : 'Edit'}`,
+            ),
+        );
     }
 
     /* --- Form --- */
@@ -100,6 +183,32 @@ export class BaseActorSheet<
         // Update document
         void this.actor.update(formData.object, { diff: false });
     }
+
+    protected async _renderFrame(
+        options: Partial<foundry.applications.api.ApplicationV2.RenderOptions>,
+    ): Promise<HTMLElement> {
+        const frame = await super._renderFrame(options);
+
+        // Insert mode toggle
+        if (this.isEditable) {
+            $(this.window.title!).before(`
+                <label id="mode-toggle" 
+                    class="toggle-switch"
+                    data-action="toggle-mode"
+                    data-tooltip="COSMERE.Actor.Sheet.Edit"
+                >
+                    <input type="checkbox" ${this.mode === 'edit' ? 'checked' : ''}>
+                    <div class="slider rounded">
+                        <i class="fa-solid fa-pen"></i>
+                    </div>
+                </label>
+            `);
+        }
+
+        return frame;
+    }
+
+    /* --- Lifecycle --- */
 
     /* --- Context --- */
