+++ src/system/applications/actor/components/character/index.ts
@@ -1,16 +1,9 @@
-export * from './details';
-export * from './resource';
-export * from './attributes';
 export * from './skills-group';
 export * from './expertises';
 export * from './ancestry';
 export * from './paths';
-export * from './actions-list';
-export * from './search-bar';
-export * from './equipment-list';
+export * from '../equipment-list';
 export * from './goals-list';
 export * from './connections-list';
-export * from './conditions';
-export * from './injuries-list';
-export * from './effects-list';
+export * from '../effects-list';
 export * from './favorites';
