+++ src/system/applications/actor/components/index.ts
@@ -1,0 +1,7 @@
+export * from './details';
+export * from './resource';
+export * from './attributes';
+export * from './actions-list';
+export * from './search-bar';
+export * from './conditions';
+export * from './injuries-list';
