+++ src/templates/actors/components/attributes.hbs
