+++ src/templates/actors/components/conditions.hbs
