+++ src/templates/actors/components/details.hbs
