+++ src/templates/actors/components/injuries-list.hbs
