+++ src/templates/actors/components/resource.hbs
