+++ src/templates/actors/components/search-bar.hbs
